requirements.txt
# Core
Django==5.2.3
psycopg2-binary==2.9.10
gunicorn==23.0.0

# REST API
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-filter==25.1

# Security & CORS
django-cors-headers==4.7.0

# Celery
celery==5.3.4
django-celery-beat==2.8.1

# Redis (совместимая версия для Django 5.2)
redis>=5.0.1,<6.0

# Utils
python-dotenv==1.1.0
Pillow==11.3.0
stripe==12.5.0
django-timezone-field==7.1

# Docs
drf-yasg==1.21.10